proteinscope/requirements.txt
# Core web framework
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
python-multipart>=0.0.9

# HTTP client
httpx>=0.27,<1.0

# Data validation
pydantic>=2.5,<3.0

# CLI
click>=8.1,<9.0

# Environment
python-dotenv>=1.0,<2.0

# XML / HTML parsing
lxml>=5.0,<6.0
xmltodict>=0.13,<1.0

# PDF generation
reportlab>=4.1,<5.0

# Image handling
Pillow>=10.0,<12.0

# Templating
jinja2>=3.1,<4.0

# Markdown rendering
markdown>=3.6,<4.0

# Bioinformatics
biopython>=1.83,<2.0
pyhmmer>=0.10,<1.0
biotite>=0.38,<1.0

# PyTorch — CPU version pinned here; for CUDA replace with the appropriate
# torch+cu* wheel from https://download.pytorch.org/whl/
# Example CUDA 12.4:  pip install torch --index-url https://download.pytorch.org/whl/cu124
torch>=2.2,<3.0

# ML / embeddings / similarity search
sentence-transformers>=3.0,<4.0
# faiss-cpu is the default.  On CUDA hosts replace with faiss-gpu:
#   pip install faiss-gpu
faiss-cpu>=1.8,<2.0

# ESM-2 protein language model (required by variant_scorer.py and inverse_folder.py)
# Pin to 2.0.0 for stability; if install fails try: pip install fair-esm --no-deps
fair-esm==2.0.0

# PROPKA — pKa prediction for metalloenzyme analysis (P4)
propka>=3.5.0

# ProDy — ANM/GNM conformational analysis (Phase 2)
ProDy>=2.4.0

# Graph analysis
networkx>=3.3,<4.0

# Google Gemini AI
google-generativeai>=0.7,<1.0

# SOAP / BRENDA
zeep>=4.2,<5.0

# PDF (alternative renderer — optional but listed for completeness)
weasyprint>=62.0,<63.0

# Testing
pytest>=8.0,<9.0
pytest-httpx>=0.30,<1.0
pytest-asyncio>=0.23,<1.0
